requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
requests==2.32.4
rsa==4.9.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.9.0
